ENH: added 'groups' functionality to tasks lists.
The task-list (and 'Task' object) can now have an additional 'groups' parameter, which accepts a list of strings.
There is a corresponding command-line argument '--task-groups' which also accepts a (list of) string(s).  These groups
are cross-references, and tasks in matching groups are enabled.
The behavior of the previous '--tasks' argument is also changed slightly: it now disables all tasks, then enables those
listed.  This can now be combined with additional task flags to start with no tasks, then add them in different ways.