ENH: change from 'entry_to_filename' as a util method to 'Entry.get_filename'.
Call signature is the same: 'get_filename(name) -> str filename'.
Implemented as a class method so it can be called without an instance (e.g. for 'Entry.init_from_file').  Having this as a
class method allows for subclassing as desired.  Seems intuitive to have as an  method as the type of entry should
determine what rules are appropriate for converting from name to filename.